Change example command to delete Docker container
Adding `--rm` tells Docker to delete the container it creates from the
`latex` image after it has completed its work. This prevents a bunch of
zombie containers from remaining on the host system, one from each time
the sample command is executed.

If needed, the `--rm` flag could be removed in order to preserve the
container (and its file system) after execution, if needed for debug
purposes.

https://docs.docker.com/reference/cli/docker/container/run/#rm